added table check, fix mapper, adjust filesplugin

+++ lib/Migration/Version1003Date202311061844.php
@@ -146,8 +146,10 @@ class Version1003Date202311061844 extends SimpleMigrationStep {
 			$table->addUniqueIndex(['appid', 'type', 'name', 'path'], 'ui_style__idx');
 		}
 
-		$table = $schema->getTable('ex_files_actions_menu');
-		$table->dropColumn('icon_class');
+		if ($schema->hasTable('ex_files_actions_menu')) {
+			$table = $schema->getTable('ex_files_actions_menu');
+			$table->dropColumn('icon_class');
+		}
 
 		return $schema;
 	}
